Fixed a few bugs with user-specific content asset management; updated to latest Snow compiler

+++ src/lib/lang_data.php
@@ -104,13 +104,13 @@ class LangData
 
 	function LangData($str_lang)
 	{
-		$this->obj_sql = new Database("tbl_prailsbase_");
-		$this->setLanguage($str_lang);
-		$this->arr_item_cache = Array();
 		if (IS_SETUP)
 			$this->uid = if_set($_SESSION['builder']['user_id'], crc32("devel"));
 		else
 			$this->uid = -1;
+		$this->obj_sql = new Database("tbl_prailsbase_");
+		$this->setLanguage($str_lang);
+		$this->arr_item_cache = Array();
 	}
 	
         /**
@@ -222,7 +222,7 @@ class LangData
 	 * get a list of all content assets across all languages
 	 **/
 	function listTexts() {
-		$arr_result = $this->obj_sql->query("SELECT a.* FROM tbl_prailsbase_texts AS a, tbl_prailsbase_language AS b WHERE a.fk_language_id=b.language_id AND a.fk_user_id=".$this->uid." GROUP BY a.identifier, a.texts_id, a.fk_language_id, a.content");
+		$arr_result = $this->obj_sql->query("SELECT a.* FROM tbl_prailsbase_texts AS a, tbl_prailsbase_language AS b WHERE a.fk_language_id=b.language_id AND b.fk_user_id=".$this->uid." GROUP BY a.identifier, a.texts_id, a.fk_language_id, a.content");
 		$arr_return = Array();
 		foreach ($arr_result as $arr_entry) {
 			$parts = explode(".", $arr_entry["identifier"]);
@@ -247,7 +247,7 @@ class LangData
 	 * folder to be listed for a menu or create an order on those.
 	 **/
 	function listAllTextsFromRoot($rootNode) {
-		$arr_result = $this->obj_sql->query("SELECT a.* FROM tbl_prailsbase_texts AS a, tbl_prailsbase_language AS b WHERE a.fk_language_id=b.language_id AND a.fk_user_id=".$this->uid." AND a.identifier LIKE '".$rootNode.".%'");
+		$arr_result = $this->obj_sql->query("SELECT a.* FROM tbl_prailsbase_texts AS a, tbl_prailsbase_language AS b WHERE a.fk_language_id=b.language_id AND b.fk_user_id=".$this->uid." AND a.identifier LIKE '".$rootNode.".%'");
 		foreach($arr_result as &$item) {
 			$item["custom"] = $this->_decodeCustom($item["custom"]);
 		}
@@ -261,7 +261,7 @@ class LangData
 	 * Returns all content assets that contain the keyword in some way - regardless of active language.
 	 **/
 	function findTextByContent($word) {
-		$arr_result = $this->obj_sql->query("SELECT a.* FROM tbl_prailsbase_texts AS a, tbl_prailsbase_language AS b WHERE a.fk_language_id=b.language_id AND a.fk_user_id=".$this->uid." AND content LIKE '%".$word."%'");
+		$arr_result = $this->obj_sql->query("SELECT a.* FROM tbl_prailsbase_texts AS a, tbl_prailsbase_language AS b WHERE a.fk_language_id=b.language_id AND b.fk_user_id=".$this->uid." AND content LIKE '%".$word."%'");
 		$arr_return = Array();
 		foreach ($arr_result as $res) {
 			array_push($arr_return, Array("id" => "text_.".$res['identifier'], "name" => $res["identifier"], "type" => "text", "custom" => $this->_decodeCustom($res["custom"])));
@@ -276,7 +276,7 @@ class LangData
 	 * This method fetches all content assets across all languages it exists in that have the given identifier.
 	 **/
 	function getAllTextsByIdentifier($ident) {
-		$texts = $this->obj_sql->query("SELECT * FROM tbl_prailsbase_language AS b LEFT JOIN tbl_prailsbase_texts AS a ON identifier='".$ident."' AND b.language_id=a.fk_language_id AND b.fk_user_id=".$this->uid." WHERE 1=1");
+		$texts = $this->obj_sql->query("SELECT * FROM tbl_prailsbase_language AS b LEFT JOIN tbl_prailsbase_texts AS a ON identifier='".$ident."' AND b.language_id=a.fk_language_id WHERE b.fk_user_id=".$this->uid."");
 		foreach ($texts as &$text) {
 			$text["default"] = $text["isDefault"];
 			$text["custom"] = $this->_decodeCustom($text["custom"]);
@@ -351,14 +351,14 @@ class LangData
 	function updateLanguage($id, $arr_data) {
 		$arr_data["isDefault"] = $arr_data["default"];
 //		$arr_languages = $this->listLanguages();
-        $this->obj_sql->update(tbl_prailsbase_language, Array("isDefault" => "0"), "1=1");
+        $this->obj_sql->update(tbl_prailsbase_language, Array("isDefault" => "0"), "fk_user_id=".$this->uid."");
 /*
         foreach ($arr_languages as $lang) {
 			if ($lang["isDefault"] == 0 && $id == $lang["language_id"]) {
 				$arr_data["isDefault"] = 1;
 			}
 		}//*/
-		$this->obj_sql->update(tbl_prailsbase_language, $arr_data, "language_id=".(int)$id."");
+		$this->obj_sql->update(tbl_prailsbase_language, $arr_data, "language_id=".(int)$id." AND fk_user_id=".$this->uid."");
 	}
 
 	function insertLanguage($arr_data) {
@@ -370,7 +370,7 @@ class LangData
       if ($arr_data["isDefault"] == 1 && $count > 0) {
           // if the to-be-created language should be the new default, then 
           // set all other default languages to non-default
-          $this->obj_sql->update(tbl_prailsbase_language, Array("isDefault" => "0"), "1=1");
+          $this->obj_sql->update(tbl_prailsbase_language, Array("isDefault" => "0"), "fk_user_id=".$this->uid."");
       }
       $id = $this->obj_sql->add(tbl_prailsbase_language, $arr_data);
       return $id;
@@ -378,11 +378,11 @@ class LangData
    
    function deleteLanguage($id) {
       $this->obj_sql->delete(tbl_prailsbase_texts, "fk_language_id=".(int)$id." OR fk_language_id=0");      
-   	  $this->obj_sql->delete(tbl_prailsbase_language, "language_id=".(int)$id."");
+   	  $this->obj_sql->delete(tbl_prailsbase_language, "language_id=".(int)$id." AND fk_user_id=".$this->uid."");
    }
 
    function deleteLanguageOnly($id) {
-      $this->obj_sql->delete(tbl_prailsbase_language, "language_id=".(int)$id."");
+      $this->obj_sql->delete(tbl_prailsbase_language, "language_id=".(int)$id." AND fk_user_id=".$this->uid."");
    }
    
    function _encodeCustom($data) {
